# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a k8s management platform API documentation",
        "title": "K8s Platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/dict/type": {
            "put": {
                "description": "根据ID更新字典类型信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典类型管理"
                ],
                "summary": "更新字典类型",
                "parameters": [
                    {
                        "description": "字典类型保存请求参数",
                        "name": "dictTypeSaveRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DictTypeSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的字典类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典类型管理"
                ],
                "summary": "创建字典类型",
                "parameters": [
                    {
                        "description": "字典类型保存请求参数",
                        "name": "dictTypeSaveRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DictTypeSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/dict/type/page": {
            "post": {
                "description": "根据条件获取字典类型的分页数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典类型管理"
                ],
                "summary": "获取字典类型分页列表",
                "parameters": [
                    {
                        "description": "字典类型分页请求参数",
                        "name": "dictTypePageRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DictTypePageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/dict/type/{id}": {
            "get": {
                "description": "根据ID获取字典类型的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典类型管理"
                ],
                "summary": "获取字典类型详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除字典类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典类型管理"
                ],
                "summary": "删除字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/dict/type/{id}/{status}": {
            "put": {
                "description": "根据ID更新字典类型的状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "字典类型管理"
                ],
                "summary": "更新字典类型状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "状态值",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/jobs/plan/create": {
            "post": {
                "tags": [
                    "作业计划管理"
                ],
                "summary": "创建计划",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JobPlanSaveRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/jobs/plan/delete": {
            "delete": {
                "tags": [
                    "作业计划管理"
                ],
                "summary": "删除计划",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "计划ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/jobs/plan/detail": {
            "get": {
                "tags": [
                    "作业计划管理"
                ],
                "summary": "获取计划详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "计划ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/jobs/plan/page": {
            "post": {
                "tags": [
                    "作业计划管理"
                ],
                "summary": "获取计划分页",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JobPlanPageRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/jobs/plan/update": {
            "post": {
                "tags": [
                    "作业计划管理"
                ],
                "summary": "更新计划",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JobPlanSaveRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/jobs/schedule/create": {
            "post": {
                "tags": [
                    "定时任务管理"
                ],
                "summary": "创建任务",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JobScheduledTaskSaveRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/jobs/schedule/delete": {
            "delete": {
                "tags": [
                    "定时任务管理"
                ],
                "summary": "删除任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/jobs/schedule/page": {
            "post": {
                "tags": [
                    "定时任务管理"
                ],
                "summary": "获取任务分页",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JobScheduledTaskPageRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/jobs/schedule/update": {
            "post": {
                "tags": [
                    "定时任务管理"
                ],
                "summary": "更新任务",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.JobScheduledTaskSaveRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "dto.DictTypePageRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.DictTypeSaveRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.JobPlanPageRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
        "dto.JobPlanSaveRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "globalVars": {
                    "type": "string"
                },
                "hostGroupId": {
                    "type": "integer"
                },
                "hostIds": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "scriptIds": {
                    "description": "关联的脚本ID列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dto.JobScheduledTaskPageRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "dto.JobScheduledTaskSaveRequest": {
            "type": "object",
            "required": [
                "name",
                "planId",
                "strategy"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "planId": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "strategy": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}